+++ src_ts/components/pages/interventions/intervention-tab-pages/intervention-details/document-details/documentDetails.selectors.ts
@@ -1,7 +1,7 @@
 import {createSelector} from 'reselect';
-import {Intervention} from '../../common/models/intervention-types';
+import {Intervention, Permission, InterventionPermissionsFields} from '../../common/models/intervention-types';
 import {DocumentDetails, DocumentDetailsPermissions} from './documentDetails.models';
-import {currentIntervention, currentInterventionPermissions} from '../../selectors';
+import {currentIntervention, currentInterventionPermissions} from '../../common/selectors';
 
 export const selectDocumentDetails = createSelector(currentIntervention, (intervention: Intervention) => {
   return new DocumentDetails(intervention);
@@ -9,7 +9,7 @@ export const selectDocumentDetails = createSelector(currentIntervention, (interv
 
 export const selectDocumentDetailsPermissions = createSelector(
   currentInterventionPermissions,
-  (permissions: Intervention) => {
+  (permissions: Permission<InterventionPermissionsFields>) => {
     return {
       edit: new DocumentDetailsPermissions(permissions!.edit),
       required: new DocumentDetailsPermissions(permissions!.required)
