+++ src_ts/components/pages/interventions/intervention-tab-pages/intervention-details/intervention-details.ts
@@ -1,4 +1,4 @@
-import {LitElement, html, property} from 'lit-element';
+import {LitElement, html} from 'lit-element';
 import {connect} from '../utils/store-subscribe-mixin';
 import './partner-details/partner-details';
 import './document-details/document-details';
@@ -13,9 +13,7 @@ export class InterventionDetails extends connect(LitElement) {
     // language=HTML
     return html`
       ${gridLayoutStylesLit}
-      <style>
-       
-      </style>
+      <style></style>
 
       <partner-details></partner-details>
       <unicef-details></unicef-details>
