+++ src_ts/components/pages/interventions/intervention-tab-pages/intervention-details/partner-details/partnerDetails.selectors.ts
@@ -1,7 +1,7 @@
 import {createSelector} from 'reselect';
 import {PartnerDetails, PartnerDetailsPermissions} from './partnerDetails.models';
-import {Intervention} from '../../common/models/intervention-types';
-import {currentIntervention, currentInterventionPermissions} from '../../selectors';
+import {Intervention, Permission, InterventionPermissionsFields} from '../../common/models/intervention-types';
+import {currentIntervention, currentInterventionPermissions} from '../../common/selectors';
 
 export const selectPartnerDetails = createSelector(currentIntervention, (intervention: Intervention) => {
   return new PartnerDetails(intervention);
@@ -9,7 +9,7 @@ export const selectPartnerDetails = createSelector(currentIntervention, (interve
 
 export const selectPartnerDetailsPermissions = createSelector(
   currentInterventionPermissions,
-  (permissions: Intervention) => {
+  (permissions: Permission<InterventionPermissionsFields>) => {
     return {
       edit: new PartnerDetailsPermissions(permissions!.edit),
       required: new PartnerDetailsPermissions(permissions!.required)
